feature: finish simple nutrition card design
